extract http logic into a helper
