+++ src/main/java/org/cru/redegg/reporting/common/Payload.java
@@ -1,0 +1,13 @@
+package org.cru.redegg.reporting.common;
+
+import java.io.IOException;
+import java.io.Writer;
+
+/**
+ * @author Matt Drees
+ */
+public interface Payload
+{
+
+    void writeTo(Writer writer) throws IOException;
+}
