+++ src/main/java/org/cru/redegg/reporting/errbit/NativeErrbitReporter.java
@@ -1,24 +1,12 @@
 package org.cru.redegg.reporting.errbit;
 
-import com.google.common.base.Charsets;
-import com.google.common.io.CharStreams;
-import com.google.common.io.Closeables;
 import org.apache.log4j.Logger;
 import org.cru.redegg.reporting.ErrorReport;
 import org.cru.redegg.reporting.api.ErrorReporter;
+import org.cru.redegg.reporting.common.HttpPayloadSender;
 
 import javax.inject.Inject;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.io.Reader;
-import java.io.Writer;
-import java.net.HttpURLConnection;
-import java.net.ProtocolException;
-import java.net.URI;
-import java.util.concurrent.TimeUnit;
 
 /**
  * Reports errors to an Errbit instance using the v2.4 xml api.
@@ -59,12 +47,13 @@ public class NativeErrbitReporter implements ErrorReporter
         log.info("user error: " + report.getRootErrorMessage().or("<message not available>"));
     }
 
+
     private void doSend(ErrorReport report)
     {
         ErrbitXmlPayload payload = new ErrbitXmlPayload(report, config);
         try
         {
-            sendXmlReport(payload);
+            new HttpPayloadSender(config.getEndpoint(), "application/xml").send(payload);
         }
         catch (IOException e)
         {
@@ -72,81 +61,4 @@ public class NativeErrbitReporter implements ErrorReporter
         }
     }
 
-    private int sendXmlReport(ErrbitXmlPayload payload) throws IOException
-    {
-        HttpURLConnection urlConnection = buildConnection();
-        configure(urlConnection);
-        sendXmlPayload(payload, urlConnection);
-        int responseCode = urlConnection.getResponseCode();
-        if (responseCode >= 400)
-        {
-            throw new RuntimeException(
-                "notice not successfully submitted; response code: " + responseCode +
-                "; content:\n" + getContent(urlConnection));
-        }
-
-        return responseCode;
-    }
-
-    private String getContent(HttpURLConnection urlConnection) throws IOException
-    {
-        InputStream errorStream = urlConnection.getErrorStream();
-        if (errorStream == null)
-            return "(no content)";
-        else
-        {
-            return readErrorStream(errorStream);
-        }
-    }
-
-    private String readErrorStream(InputStream errorStream) throws IOException
-    {
-        Reader reader = new InputStreamReader(errorStream, Charsets.UTF_8);
-        boolean threw = true;
-        try
-        {
-            String content = CharStreams.toString(reader);
-            threw = false;
-            return content;
-        }
-        finally
-        {
-            Closeables.close(reader, threw);
-        }
-    }
-
-    private void sendXmlPayload(ErrbitXmlPayload payload, HttpURLConnection urlConnection) throws IOException
-    {
-        urlConnection.connect();
-        OutputStream outputStream = urlConnection.getOutputStream();
-        Writer writer = new OutputStreamWriter(outputStream, Charsets.UTF_8);
-        boolean threw = true;
-        try
-        {
-            payload.writeXmlTo(writer);
-            threw = false;
-        }
-        finally
-        {
-            Closeables.close(writer, threw);
-        }
-    }
-
-    private void configure(HttpURLConnection urlConnection) throws ProtocolException
-    {
-        urlConnection.setDoOutput(true);
-        urlConnection.setRequestMethod("POST");
-        urlConnection.setRequestProperty("Content-Type", "application/xml");
-        urlConnection.setConnectTimeout((int) TimeUnit.SECONDS.toMillis(10));
-        urlConnection.setReadTimeout((int) TimeUnit.SECONDS.toMillis(30));
-    }
-
-    private HttpURLConnection buildConnection() throws IOException
-    {
-        URI endpoint = config.getEndpoint();
-        if (endpoint == null)
-            throw new IllegalArgumentException("no endpoint is configured!");
-
-        return (HttpURLConnection) endpoint.toURL().openConnection();
-    }
 }
